+++ opentelemetry-appender-tracing/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## vNext
 
-- Populate TimeStamp on LogRecord [#1175](https://github.com/open-telemetry/opentelemetry-rust/pull/1175).
+- Add log appender versions to loggers (#1182)
 
 ## v0.1.0