+++ opentelemetry-proto/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Implement tonic metrics proto transformations (#1184)
 
+### Fixed
+
+- Rename `traces` feature to the more standard `trace` (#1183)
+
 ## v0.3.0
 
 ### Updated